audio/requirements.txt

chatterbox-tts==0.1.4
clip-anytorch==2.6.0
conformer==0.3.2
curated-transformers==0.1.1
dctorch==0.1.2
diffusers==0.29.0
gradio==5.44.1
gradio_client==1.12.1
gradio_logsview @ https://huggingface.co/spaces/cocktailpeanut/gradio_logsview/resolve/main/gradio_logsview-0.0.17-py3-none-any.whl#sha256=c595d1ab65cc04b009b538652f64dcdb2e9d30491ee0ffc12f25101da5e77042
lion-pytorch==0.2.3
numpy==1.25.2
open_clip_torch==3.1.0
python-dotenv==1.2.1
pytorch-fid==0.3.0
pytorch_optimizer==3.7.0
resemble-perth==1.0.1
s3tokenizer==0.2.0
scipy==1.15.3
spacy-curated-transformers==0.3.1
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.1
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.24.0
transformers==4.46.3

